changed amount to int instead of float to use synchronization primitives and changed client.deposit to use compare and swap

+++ bank.h
@@ -7,6 +7,7 @@
  * 12/31/19    Tim Liu    removed transaction struct - moved to bank_types.h
  *                       to avoid circular dependency
  * 12/31/19    Tim Liu    added mutex bank_lock as private member 
+ * 01/02/20    Tim Liu    changed interest and balance from float to int
  *
  */
 
@@ -32,31 +33,31 @@ class Bank {
     map<string, Client> client_list;       // list of clients - super duper private!!
     //queue<transaction> trans_queue;        // queue of pending transactions for bank to handle
     mQueue trans_queue;                    // use multiqueue for transactions
-    float interest;                        // bank interest rate as a percent
+    int interest;                        // bank interest rate as a percent
     mutex bank_lock;                       // lock for the transaction queue
 
 public:
     // constructor - by default create a bank with 1.0% interest
-    Bank(float interest=1.0f);
+    Bank(int interest=1);
     // destructor
     ~Bank();
 
     // summary functions
-    float get_total_balance();        // calculate total bank holdings
-    float get_mean();                 // get average balance
+    int get_total_balance();        // calculate total bank holdings
+    int get_mean();                 // get average balance
     void show_clients();              // print all clients and balances
 
     // transactions
-    void new_client(const string &name, float balance=0.0f);      // add a new client
-    void bank_deposit(const string &client_name, float amount);   // deposit money into a client's account
-    void bank_withdraw(const string &client_name, float amount);  // withdraw money from a client account
+    void new_client(const string &name, int balance=0);      // add a new client
+    void bank_deposit(const string &client_name, int amount);   // deposit money into a client's account
+    void bank_withdraw(const string &client_name, int amount);  // withdraw money from a client account
 
-    void bank_transfer(const string &from_client, const string &to_client, float amount);   // transfer money
+    void bank_transfer(const string &from_client, const string &to_client, int amount);   // transfer money
     void bank_pay_interest();                                     // add interest payment to all accounts
 
     // functions for handling transactions
     void handle_transactions();                           // performs transactions in trans_queue
-    void add_transaction(int type, float amount, const string &client_a, const string &client_b); 
+    void add_transaction(int type, int amount, const string &client_a, const string &client_b); 
 
 };
